% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{amazonForestMask}
\alias{amazonForestMask}
\title{Apply amazon forest mask}
\usage{
amazonForestMask(myraster, legend)
}
\arguments{
\item{myraster}{A raster data.}

\item{legend}{A tibble legend.}
}
\description{
Apply amazon forest mask.
}
