% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{amazonMask}
\alias{amazonMask}
\title{Apply amazon mask}
\usage{
amazonMask(myraster)
}
\arguments{
\item{myraster}{A raster data.}
}
\description{
Apply amazon biome mask in order to remove all pixels that do not
belong to the biome
}
