% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{amazonWaterMask}
\alias{amazonWaterMask}
\title{Apply amazon water mask}
\usage{
amazonWaterMask(myraster, legend)
}
\arguments{
\item{myraster}{A raster.}

\item{legend}{A tibble legend.}
}
\description{
Apply amazon water mask from Perkel et al. (doi:10.1038/nature20584) over a raster, given a legend and an year.
}
