% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{applyFunctionByBlocks}
\alias{applyFunctionByBlocks}
\title{Apply a function over all pixels of a image by block}
\usage{
applyFunctionByBlocks(file, block_func)
}
\arguments{
\item{file}{A tiff file.}

\item{block_func}{A function that takes a vector of pixels as argument
and returns another vector of pixels with the same size of the input.
All but the last call use vectors with the same size. The last one
uses the remaining pixels of the raster.}
}
\description{
Generate a raster from another raster an a function that
takes block (a vector of pixels) as argument and returns a new raster
in memory.
}
