% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{baseDir}
\alias{baseDir}
\title{Base path}
\usage{
baseDir(path)
}
\arguments{
\item{path}{A path that within the base directory.}
}
\description{
Return a path according to the base directory of the package.
}
\seealso{
setBaseDir
}
