% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{buildStyle}
\alias{buildStyle}
\title{Create a style file}
\usage{
buildStyle(legend, outputFile)
}
\arguments{
\item{legend}{A tibble with four columns as described in readLegend().}

\item{outputFile}{A file name to be created. It tipically has extension .qml.}
}
\description{
Create a QGIS style file from a legend.
}
