% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{cerradoMask}
\alias{cerradoMask}
\title{Apply cerrado mask}
\usage{
cerradoMask(myraster)
}
\arguments{
\item{myraster}{A raster data.}
}
\description{
Apply cerrado biome mask in order to remove all pixels that do not
belong to the biome
}
