% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{cerradoUrbanMask}
\alias{cerradoUrbanMask}
\title{Apply cerrado urban mask}
\usage{
cerradoUrbanMask(myraster, legend, year)
}
\arguments{
\item{myraster}{A raster.}

\item{legend}{A tibble legend.}

\item{year}{An year. Sugarcane data are available for 2000 to 2016.}
}
\description{
Apply cerrado urban mask from ?? over a raster, given a legend and an year.
}
