% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{cerradoWaterMask}
\alias{cerradoWaterMask}
\title{Apply cerrado water mask}
\usage{
cerradoWaterMask(myraster, legend)
}
\arguments{
\item{myraster}{A raster.}

\item{legend}{A tibble legend.}
}
\description{
Apply cerrado water mask from Perkel et al. (doi:10.1038/nature20584) over a raster, given a legend and an year.
}
