% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{compareRasters}
\alias{compareRasters}
\title{Compare two categorical rasters}
\usage{
compareRasters(data, reference)
}
\arguments{
\item{data}{String with the input data file.}

\item{reference}{String with the reference data file.}
}
\description{
Compares two categorical rasters with the same projection, extent, and resolution.
It returns a matrix with the percentage of the area covered by each pair of values from the
two rasters.
}
