% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{compareRasters_area}
\alias{compareRasters_area}
\title{Compare two categorical rasters weitghing each class by its area.}
\usage{
compareRasters_area(data, reference)
}
\arguments{
\item{data}{String with the input data file.}

\item{reference}{String with the reference data file.}
}
\description{
Compares two categorical rasters with the same projection, 
extent, and resolution. It returns a list made of a confusion matrix, a named 
vector of areas (in the units of the reference raster), a sublist of vectors 
of the superior and inferior confidence intervals (95%), and a sublist of 
vectors of the user and producer accuracies.
}
