% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTCCerrado.R
\name{compareTCCerrado}
\alias{compareTCCerrado}
\title{Compare with TerraClass Cerrado}
\usage{
compareTCCerrado(data, progress = TRUE)
}
\arguments{
\item{data}{A raster data.}

\item{progress}{A boolean value indicating whether this function should print its progress. Default is true.}
}
\description{
Compares the data with TerraClass Cerrado data for 2013. It returns values in hectares.
}
