% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{countPixels}
\alias{countPixels}
\title{Count pixels of each class}
\usage{
countPixels(filename)
}
\arguments{
\item{filename}{String with the file name.}
}
\description{
Count the number of pixels of each class in a raster file.
}
