% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createEmptyDir}
\alias{createEmptyDir}
\title{Create an empty directory}
\usage{
createEmptyDir(directory)
}
\arguments{
\item{directory}{A relative path within baseDir().}
}
\description{
Create an empty directory or clean an existent one.
}
