% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degreesToMeters}
\alias{degreesToMeters}
\title{Convert degrees to meters}
\usage{
degreesToMeters(degrees)
}
\arguments{
\item{degrees}{A value in degrees}
}
\description{
Converts values from degrees to meters by multiplying them by 111,000.
}
