% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forEachBlock.R
\name{forEachBlock}
\alias{forEachBlock}
\title{Iterates over a raster by block}
\usage{
forEachBlock(raster, func)
}
\arguments{
\item{raster}{A raster::raster.}

\item{func}{A function that takes three arguments: (1) A vector of pixels
representing a block, (2) The row of the raster where the block starts, and
(3) The number of rows within the block.}
}
\description{
Iterates over a raster using a second order function.
}
