% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forEachBlock.R
\name{forEachBlockPair}
\alias{forEachBlockPair}
\title{Iterates over two rasters by blocks}
\usage{
forEachBlockPair(raster1, raster2, func)
}
\arguments{
\item{raster1}{A raster::raster.}

\item{raster2}{Another raster::raster.}

\item{func}{A function that takes two arguments: (1) A vector of pixels
representing a block from raster1 and (2) A vector of pixels
representing a block from raster2. Both vectors have the same size.}
}
\description{
Iterates over two rasters using a second order function. It
supposes both rasters have the same extent and resolution.
}
