% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{getEmptyRaster}
\alias{getEmptyRaster}
\title{Empty raster from a given raster file}
\usage{
getEmptyRaster(filename)
}
\arguments{
\item{filename}{Name of a file that stores raster data.}
}
\description{
Return an empty raster with the same
pixels of a given raster stored into a file.
}
