% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{getPalette}
\alias{getPalette}
\title{Get sits palette}
\usage{
getPalette(filename = baseDir("style_cerrado_13classes.qml"))
}
\arguments{
\item{filename}{File name to be read (a qml file). As default, it uses the
sits palette, stored in style_cerrado_13_classes.qml.}
}
\description{
Return the sits palette with the values, colors, and labels.
}
