% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{getSitsValidateEnv}
\alias{getSitsValidateEnv}
\title{Sits validate environment}
\usage{
getSitsValidateEnv()
}
\description{
Return the environment of sits.validate, with variables such as
base directory, and projection used internally.
}
