% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{getTifFiles}
\alias{getTifFiles}
\title{Tif files within a base directory}
\usage{
getTifFiles(dir)
}
\arguments{
\item{dir}{A directory inside base directory.}
}
\description{
Returns a vector with all tif files within a directory inside base directory.
}
