% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupToLevel.R
\name{groupDirectoryToLevel1}
\alias{groupDirectoryToLevel1}
\title{Reclassify pixels using the first level of the legend}
\usage{
groupDirectoryToLevel1(inputDir, outputDir, legend)
}
\arguments{
\item{inputDir}{Input directory. It must be a relative path within baseDir().}

\item{outputDir}{Output directory. It will be cleaned before processing
the files. It must also be a relative path within baseDir().}

\item{legend}{A legend, that can be created with readLegend().}
}
\description{
Reclassify each pixel of each image of a directory using the first
level of the legend, which are the values less than ten. The
output images are saved in the output directory.
}
