% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupToLevel.R
\name{groupRasterToLevel1}
\alias{groupRasterToLevel1}
\title{Reclassify pixels using the first level of the legend}
\usage{
groupRasterToLevel1(file, legend)
}
\arguments{
\item{file}{A tif file.}

\item{legend}{A tibble representing a legend.}
}
\description{
Reclassify each pixel of a image using the first
level of the legend, which are the values less than ten.
}
