% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{haToMha}
\alias{haToMha}
\title{Convert from ha to Mha}
\usage{
haToMha(value)
}
\arguments{
\item{value}{A vector of values.}
}
\description{
This function divides the input values
by 1e6 to convert from hectares to Millions of hectares.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
