% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{joinClassifications}
\alias{joinClassifications}
\title{Join tibles summing their areas}
\usage{
joinClassifications(classes)
}
\arguments{
\item{classes}{A list of tibbles with classifications}
}
\description{
This function joins a list of tibbles and sums
their areas. It preserves the rows and sums the values whose
columns are the same.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
