% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{landClasses}
\alias{landClasses}
\title{Creates a vector of strings from variable names to represent land use and cover classes}
\usage{
landClasses(...)
}
\arguments{
\item{...}{Names of the variables.}
}
\description{
Convert a set of variable names in a vector of strings with their names.
This function is useful to name the attributes to be read from CSV files.
}
