% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m2ToMha}
\alias{m2ToMha}
\title{Convert from m^2 to Mha}
\usage{
m2ToMha(value)
}
\arguments{
\item{value}{A vector of values.}
}
\description{
This function divides the input values
by 1e10 to convert from m^2 to Millions of hectares.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
