% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printLog}
\alias{printLog}
\title{Print a value on the screen}
\usage{
printLog(value, log)
}
\arguments{
\item{value}{The value to be printed}

\item{log}{A boolean value indicating whether value should be printed}
}
