% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printProgress}
\alias{printProgress}
\title{Print a value on the screen}
\usage{
printProgress(value, progress)
}
\arguments{
\item{value}{The value to be printed}

\item{progress}{A boolean value indicating whether value should be printed}
}
\description{
Print the progress of a function.
}
