% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rasterBoxToPolygon}
\alias{rasterBoxToPolygon}
\title{Converts a raster to a SpatialPolygons, keeping the same projection}
\usage{
rasterBoxToPolygon(raster)
}
\description{
This function returns a SpatialPolygons with the
bounding box of a raster data, with the same projection of the
original data.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
