% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{readLegend}
\alias{readLegend}
\title{Read a legend from a file}
\usage{
readLegend(legend_file = NULL)
}
\arguments{
\item{legend_file}{A csv file with the legend. As default, it uses
the sits brazil legend.}
}
\description{
Read a legend from a csv file, with four columns:
Value (with pixel values), Label (with pixel labels), Short
(with short names to be used when one wants to refer to the pixel values),
and Color (strings with hexadecimal values, such as #1e33f4).
This function returns a tibble with these four columns.
}
