% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{remapDirectory}
\alias{remapDirectory}
\title{Remap pixel values of all images within a directory}
\usage{
remapDirectory(inputDir, outputDir, replacements, overwrite = FALSE)
}
\arguments{
\item{inputDir}{Input directory. It must be a relative path within baseDir().}

\item{outputDir}{Output directory. It will be cleaned before processing
the files. It must also be a relative path within baseDir().}

\item{replacements}{A vector with the code of classes to be remapped according
to its position.}

\item{overwrite}{Should output files be overwritten?}

\item{legend}{A legend, that can be created with readLegend().}
}
\description{
create a directory with the classifications, replacing
each pixel of each image by a given set of values.
Every value in the original image must be mapped, even
if the updated value is the same.
}
\details{
This function also creates a qml file with the colors
of the pixels.
}
\seealso{
remapRaster()
}
