% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{reSample}
\alias{reSample}
\title{Resample raster data}
\usage{
reSample(inputFile, outputFile, reference)
}
\arguments{
\item{inputFile}{File name to be resampled.}

\item{outputFile}{File name to be created.}

\item{reference}{The reference data whose resolution and extent will be
used for the otput file.}
}
\description{
Resample a raster file using the resolution and extent of a
reference data.
}
