% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyOutput}
\alias{simplifyOutput}
\title{Simplify the comparison output}
\usage{
simplifyOutput(output)
}
\arguments{
\item{output}{The output of a compare function, such as compareTCCerrado() or compareCerradoMask().}
}
\description{
This function returns a simplified version of
a comparison output. It removes all rows and columns with only
zeros as values. It also adds a new column and a new row with
totals for the respective column and row. Note that if you use
this function twice with the same data, it will consider create
a new column and row with the totals.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
