% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitRaster}
\alias{splitRaster}
\title{Split a raster file}
\usage{
splitRaster(inputFile, outputDir, n.side)
}
\arguments{
\item{inputFile}{Name of the input file}

\item{outputDir}{Output directory where the files will be saved.}

\item{n.side}{Number of splits in each side (x and y).}
}
\description{
Split a raster file into n.side ^ 2 files. Each output file
will be have the same name of the input file plut -tile-<number>.
}
\seealso{
splitRasters
}
