% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{subLegend}
\alias{subLegend}
\title{Compute a sublegend}
\usage{
subLegend(legend, myraster)
}
\arguments{
\item{legend}{A Legend, which might be read from readLegend(), or also a sublegend
created from this function.}

\item{myraster}{A raster or the path to a raster file.}
}
\description{
Compute a sublegend based on a legend
the values available in a raster.
}
