% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCerradoMask.R
\name{summarizeAsPercentage}
\alias{summarizeAsPercentage}
\title{Summarize the results as percentages}
\usage{
summarizeAsPercentage(result)
}
\arguments{
\item{result}{A classification result.}
}
\description{
Summarize the results for a given region using percentages
instead of the absolute values.
}
