% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeByPolygons.R
\name{summarizeOneByMunicipalities}
\alias{summarizeOneByMunicipalities}
\title{Summarize the classification areas in each Brazilian municipality.}
\usage{
summarizeOneByMunicipalities(data, file = "55mu2500gsd_removednull",
  name = "Nome_Munic", progress = TRUE)
}
\arguments{
\item{data}{A raster::raster data.}

\item{file}{Name of a shapefile that store polygons to be used to summarize the data.}

\item{name}{String with the attribute that stores names of the municipalities.}

\item{progress}{A boolean value indicating whether this function should print its progress. Default is true.}
}
\description{
This function returns the area of each class within
each Brazilian municipality. The output is a data.frame with the
same row names (classes) and municipalities as columns. It uses file
Dropbox/sits.validate/shapes/55mu2500gsd_removednull.shp, with 5,564 municipalities.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
