% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeByPolygons.R
\name{summarizeOneByPolygons}
\alias{summarizeOneByPolygons}
\title{Summarize the classification areas into a set of polygons}
\usage{
summarizeOneByPolygons(data, layer, attribute, progress = TRUE)
}
\arguments{
\item{data}{A raster::raster data.}

\item{layer}{Name of the layer with the polygons.}

\item{attribute}{Name of an attribute from the polygons data to be shown in the progress.}

\item{progress}{A boolean value indicating whether this function should print its progress. Default is true.}
}
\description{
This function returns a table separating the
classifications according to the overlapping areas with a
set of polygons.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
