% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCerradoMask.R
\name{totalValidationCerradoMask}
\alias{totalValidationCerradoMask}
\title{Computes the total agreement for cerrado mask}
\usage{
totalValidationCerradoMask(result)
}
\arguments{
\item{result}{A classification result.}
}
\description{
The total agreement for cerrado mask includes
}
