% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTCCerrado.R
\name{totalValidationTCCerrado}
\alias{totalValidationTCCerrado}
\title{Return the total validation comparing with TerraClass Cerrado}
\usage{
totalValidationTCCerrado(result)
}
\arguments{
\item{result}{The result of compareTCCerrado().}
}
\description{
The total validation is the sum of Soy_Corn, Soy_Cotton, Soy_Fallow, Soy_Millet, Sugarcane areas
classified as Agricultura anual, plus Araguaia, Campo_Cerrado, Cerradao, Cerrado, and Cerrado_Rupestre areas
classified as Natural, plus Pasture areas classified as Pastagem.
}
\seealso{
compareTCCerrado
}
