% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{waterMask}
\alias{waterMask}
\title{Applies water mask}
\usage{
waterMask(myraster, legend)
}
\arguments{
\item{myraster}{A raster.}

\item{legend}{A tibble with the legend, as described by readLegend().}
}
\description{
Updates the pixels that overlap the water pixels
into water. This function only works for Brazil.
}
