% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{baseDir}
\alias{baseDir}
\title{Get or set base directory}
\usage{
baseDir(dir = NULL)
}
\arguments{
\item{dir}{A directory.}
}
\description{
Set the base directory (if the argument is used).
It returns the current base directory.
All the data used internally by the package are stored in this path. It is a
Dropbox folder that cannot be stored in GitHub due to its size. If you do not
have access to this folder please contact the package's maintainer.
}
