% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTCCerrado.R
\name{compareTCCerrado}
\alias{compareTCCerrado}
\title{Compare with TerraClass Cerrado}
\usage{
compareTCCerrado(data, log = TRUE)
}
\arguments{
\item{data}{A raster data.}

\item{log}{A boolean value indicating whether a log with the processing steps should be printed in the screen.
Default is true.}
}
\description{
Compares the data with TerraClass Cerrado data for 2013. It returns values in hectares.
}
