% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forEachBlock.R
\name{forEachBlock}
\alias{forEachBlock}
\title{Iterates over a raster by block}
\usage{
forEachBlock(raster, func, ...)
}
\arguments{
\item{raster}{A raster::raster.}

\item{func}{A function that takes at least one arguments: A vector of pixels
representing a block.}

\item{...}{Any other parameters to pass to func.}
}
\description{
Iterates over a raster using a second order function.
}
