% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{landClasses}
\alias{landClasses}
\title{Creates a vector of strings from variable names to represent land use and cover classes}
\usage{
landClasses(legend, ...)
}
\arguments{
\item{legend}{A tibble legend.}

\item{...}{Names of the variables.}
}
\description{
Convert a set of class names in a vector with its code.
}
