% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{remapRaster}
\alias{remapRaster}
\title{Remap pixel values of an image}
\usage{
remapRaster(file, legend, replacements)
}
\arguments{
\item{file}{A tiff file.}

\item{legend}{A legend, that can be created with readLegend().}

\item{replacements}{A list whose indexes belong to Short values of the legend,
and whose values are the pixels of the images.}
}
\description{
Replace
each pixel of an image file by a given set of values.
Every value in the original image must be mapped, even
if the updated value is the same. It returns a raster in memory
with the updated values.
}
