% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitRasters}
\alias{splitRasters}
\title{Split all rasters within a given directory}
\usage{
splitRasters(inputDir, outputDir, n.side)
}
\arguments{
\item{outputDir}{Output directory where the files will be saved.}

\item{n.side}{Number of splits in each side (x and y).}

\item{inputFile}{Name of the input file}
}
\description{
Split all tif files within a directory, each one into n.side ^ 2 files.
}
\seealso{
splitRaster
}
