% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeByStates.R
\name{summarizeOneByMunicipalities}
\alias{summarizeOneByMunicipalities}
\alias{summarizeOneByMunicipalities}
\alias{summarizeOneByStates}
\alias{summarizeOneByMunicipalities}
\alias{summarizeOneBySimU}
\title{Summarize the classification areas in each Brazilian municipality.}
\usage{
summarizeOneByMunicipalities(data, log = TRUE)

summarizeOneByStates(data, log = TRUE)

summarizeOneBySimU(data, log = TRUE)
}
\arguments{
\item{data}{A raster::raster object.}

\item{log}{A boolean value indicating whether a log should be printed along the execution.}

\item{data}{A raster::raster object.}

\item{log}{A boolean value indicating whether a log should be printed along the execution.}
}
\description{
This function returns the area of each class within
each Brazilian municipality. The output is a data.frame with the
same row names (classes) and municipalities as columns. It uses file
55mu2500gsd_removednull.shp, with 5,564 municipalities.

This function returns the area of each class within
each Brazilian municipality. The output is a data.frame with the
same row names (classes) and municipalities as columns. It uses file
UFEBRASIL.shp, with 27 objects.

This function returns the area of each class within
each Brazilian municipality. The output is a data.frame with the
same row names (classes) and municipalities as columns. It uses file
55mu2500gsd_removednull.shp, with 18,194 simulation units.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}

Pedro R. Andrade, \email{pedro.andrade@inpe.br}

Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
