% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeByStates.R
\name{summarizeOneByPolygons}
\alias{summarizeOneByPolygons}
\title{Summarize the classification areas into a set of polygons}
\usage{
summarizeOneByPolygons(data, layer, attribute, log = TRUE)
}
\description{
This function returns a table separating the
classifications according to the overlapping areas with a
set of polygons.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
